fix(tts): update import paths after input/ → platform/darwin/ rename
