+++ tests/main/platform/darwin/selection.test.ts
@@ -2,14 +2,14 @@ import { describe, it, expect } from "vitest";
 
 describe("selection", () => {
   it("exports getSelectedText function", async () => {
-    const mod = await import("../../../src/main/input/selection");
+    const mod = await import("../../../../src/main/platform/darwin/selection");
     expect(mod.getSelectedText).toBeDefined();
     expect(typeof mod.getSelectedText).toBe("function");
   });
 
   it("returns empty string in test environment", async () => {
     const { getSelectedText } = await import(
-      "../../../src/main/input/selection"
+      "../../../../src/main/platform/darwin/selection"
     );
     const result = await getSelectedText();
     expect(result).toBe("");
